refactor folders to have a service and models folder
