+++ src/app/components/registration/registration.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { User } from '../../user';
+import { User } from './models/user';
 
 @Component({
   selector: 'app-registration',
