+++ src/app/services/registration.service.spec.ts
@@ -1,12 +1,16 @@
-import { TestBed } from '@angular/core/testing';
+import { TestBed, inject } from '@angular/core/testing';
 
 import { RegistrationService } from './registration.service';
+import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 
 describe('RegistrationService', () => {
   let service: RegistrationService;
 
   beforeEach(() => {
-    TestBed.configureTestingModule({});
+    TestBed.configureTestingModule({
+      imports: [HttpClientTestingModule],
+      
+    });
     service = TestBed.inject(RegistrationService);
   });
