+++ src/app/services/registration.service.ts
@@ -2,6 +2,8 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable, throwError } from 'rxjs';
 import { catchError, retry } from 'rxjs/operators';
+import { User } from '../components/registration/models/user';
+
 @Injectable({
   providedIn: 'root'
 })
